requirements.txt
deepgram-sdk==3.2.7
sounddevice==0.5.1
soundfile==0.12.1
numpy==1.24.3
websockets==11.0.3